Desktop View Completed

+++ src/components/Post.js
@@ -69,45 +69,64 @@ function Post(props) {
       }
       const handlePostComment = (e) => {
             e.preventDefault();
-            db
-            .collection("posts")
-            .doc(props.postId)
-            .collection("comments")
-            .add({
-                  by:props.currentUser,
-                  comment:singleComment,
-                  timestamp:firebase.firestore.FieldValue.serverTimestamp()
-            });
+            if(props.currentUser !== ""){
+                  db
+                  .collection("posts")
+                  .doc(props.postId)
+                  .collection("comments")
+                  .add({
+                        by:props.currentUser,
+                        comment:singleComment,
+                        timestamp:firebase.firestore.FieldValue.serverTimestamp()
+                  });
+            }
+            else{
+                  setOpenError(true);
+            }
             setSingleComment("");
       }
       const handleLike = (e) => {
             e.preventDefault();
-            db
-            .collection("posts")
-            .doc(props.postId)
-            .collection("likes")
-            .add({
-                  by:props.currentUser,
-                  timestamp:firebase.firestore.FieldValue.serverTimestamp()
-            });
-            setIsLiked(true);
+            if(props.currentUser !== ""){
+                  db
+                  .collection("posts")
+                  .doc(props.postId)
+                  .collection("likes")
+                  .add({
+                        by:props.currentUser,
+                        timestamp:firebase.firestore.FieldValue.serverTimestamp()
+                  });
+                  setIsLiked(true);
+            }
+            else{
+                  setOpenError(true);
+            }
       }
       const handleUnLike = (e) => {
             e.preventDefault();
-            let query = db
-            .collection("posts")
-            .doc(props.postId)
-            .collection("likes")
-            .where('by','==',props.currentUser);
-            query.get().then(function(querySnapshot) {
-              querySnapshot.forEach(function(doc) {
-                doc.ref.delete();
-              });
-            });
-            setIsLiked(false);
+            if(props.currentUser !== ""){
+                  let query = db
+                  .collection("posts")
+                  .doc(props.postId)
+                  .collection("likes")
+                  .where('by','==',props.currentUser);
+                  query.get().then(function(querySnapshot) {
+                    querySnapshot.forEach(function(doc) {
+                      doc.ref.delete();
+                    });
+                  });
+                  setIsLiked(false);
+            }
+            else{
+                  setOpenError(true);
+            }
       }
       const handleClose = () => {
           setOpenLikes(false);
+      };
+      const [OpenError, setOpenError] = new useState(false);
+      const handleCloseErrorModal = () => {
+          setOpenError(false);
       };
   return (
     <div className="Post">
@@ -131,7 +150,7 @@ function Post(props) {
                                     return(
                                           <>
                                                 <div key={like.id} className="likeListIndividualDiv">
-                                                     <img src="https://i.pinimg.com/474x/a0/4d/84/a04d849cf591c2f980548b982f461401.jpg" alt="Profile Avatar"/>
+                                                     <img src="/noProfile.jpg" alt="Profile Avatar"/>
                                                      <p>{like.data.by}</p>
                                                 </div>
                                           </>
@@ -141,9 +160,27 @@ function Post(props) {
                   </div>
             </div>
       </Modal>
+      <Modal
+      open={OpenError}
+      onClose={handleCloseErrorModal}
+      aria-labelledby="simple-modal-title"
+      aria-describedby="simple-modal-description"
+      className="Modal__Custom__OpenLikes"
+      >
+            <div className="Modal_Main_Div">
+                  <div className="likesListLabel">
+                        <CloseIcon style={{opacity : '0'}}/>
+                        <h1>Error</h1>
+                        <CloseIcon className="closeButton" onClick={handleCloseErrorModal}/>
+                  </div>
+                  <span className="NotLoginError">
+                        Login to like or comment on photos and videos of your friends.
+                  </span>
+            </div>
+      </Modal>
       <div className="header">
       	<div className="header1">
-      		<img src={(!!props.profilePic)?(props.profilePic):"https://i.pinimg.com/474x/a0/4d/84/a04d849cf591c2f980548b982f461401.jpg"} alt={`${props.by} Profile`}/>
+      		<img src={(!!props.profilePic)?(props.profilePic):"/noProfile.jpg"} alt={`${props.by} Profile`}/>
       		<div className="header1__wrapper">
       			<h1>{props.by}</h1>
       			<p>{props.location}</p>
@@ -161,7 +198,7 @@ function Post(props) {
       		<div className="iconsSide1">
       			{
                               (isLiked)?(
-                                    <svg onClick={handleUnLike} aria-label="Unlike" class="_8-yf5 " fill="#ed4956" height="24" viewBox="0 0 48 48" width="24"><path d="M34.6 3.1c-4.5 0-7.9 1.8-10.6 5.6-2.7-3.7-6.1-5.5-10.6-5.5C6 3.1 0 9.6 0 17.6c0 7.3 5.4 12 10.6 16.5.6.5 1.3 1.1 1.9 1.7l2.3 2c4.4 3.9 6.6 5.9 7.6 6.5.5.3 1.1.5 1.6.5s1.1-.2 1.6-.5c1-.6 2.8-2.2 7.8-6.8l2-1.8c.7-.6 1.3-1.2 2-1.7C42.7 29.6 48 25 48 17.6c0-8-6-14.5-13.4-14.5z"></path></svg>
+                                    <svg onClick={handleUnLike} aria-label="Unlike" className="_8-yf5 " fill="#ed4956" height="24" viewBox="0 0 48 48" width="24"><path d="M34.6 3.1c-4.5 0-7.9 1.8-10.6 5.6-2.7-3.7-6.1-5.5-10.6-5.5C6 3.1 0 9.6 0 17.6c0 7.3 5.4 12 10.6 16.5.6.5 1.3 1.1 1.9 1.7l2.3 2c4.4 3.9 6.6 5.9 7.6 6.5.5.3 1.1.5 1.6.5s1.1-.2 1.6-.5c1-.6 2.8-2.2 7.8-6.8l2-1.8c.7-.6 1.3-1.2 2-1.7C42.7 29.6 48 25 48 17.6c0-8-6-14.5-13.4-14.5z"></path></svg>
                               ):(
                                     <svg onClick={handleLike} aria-label="Like" className="_8-yf5 " fill="#262626" height="24" viewBox="0 0 48 48" width="24"><path d="M34.6 6.1c5.7 0 10.4 5.2 10.4 11.5 0 6.8-5.9 11-11.5 16S25 41.3 24 41.9c-1.1-.7-4.7-4-9.5-8.3-5.7-5-11.5-9.2-11.5-16C3 11.3 7.7 6.1 13.4 6.1c4.2 0 6.5 2 8.1 4.3 1.9 2.6 2.2 3.9 2.5 3.9.3 0 .6-1.3 2.5-3.9 1.6-2.3 3.9-4.3 8.1-4.3m0-3c-4.5 0-7.9 1.8-10.6 5.6-2.7-3.7-6.1-5.5-10.6-5.5C6 3.1 0 9.6 0 17.6c0 7.3 5.4 12 10.6 16.5.6.5 1.3 1.1 1.9 1.7l2.3 2c4.4 3.9 6.6 5.9 7.6 6.5.5.3 1.1.5 1.6.5.6 0 1.1-.2 1.6-.5 1-.6 2.8-2.2 7.8-6.8l2-1.8c.7-.6 1.3-1.2 2-1.7C42.7 29.6 48 25 48 17.6c0-8-6-14.5-13.4-14.5z"></path></svg>
                               )
@@ -198,9 +235,9 @@ function Post(props) {
                   }
       	</div>
       	<div className="addComment" onKeyPress={handlePostCommentEnter}>
-      		<svg aria-label="Emoji" class="_8-yf5 " fill="#262626" height="24" viewBox="0 0 48 48" width="24"><path d="M24 48C10.8 48 0 37.2 0 24S10.8 0 24 0s24 10.8 24 24-10.8 24-24 24zm0-45C12.4 3 3 12.4 3 24s9.4 21 21 21 21-9.4 21-21S35.6 3 24 3z"></path><path d="M34.9 24c0-1.4-1.1-2.5-2.5-2.5s-2.5 1.1-2.5 2.5 1.1 2.5 2.5 2.5 2.5-1.1 2.5-2.5zm-21.8 0c0-1.4 1.1-2.5 2.5-2.5s2.5 1.1 2.5 2.5-1.1 2.5-2.5 2.5-2.5-1.1-2.5-2.5zM24 37.3c-5.2 0-8-3.5-8.2-3.7-.5-.6-.4-1.6.2-2.1.6-.5 1.6-.4 2.1.2.1.1 2.1 2.5 5.8 2.5 3.7 0 5.8-2.5 5.8-2.5.5-.6 1.5-.7 2.1-.2.6.5.7 1.5.2 2.1 0 .2-2.8 3.7-8 3.7z"></path></svg>
+      		<svg aria-label="Emoji" className="_8-yf5 " fill="#262626" height="24" viewBox="0 0 48 48" width="24"><path d="M24 48C10.8 48 0 37.2 0 24S10.8 0 24 0s24 10.8 24 24-10.8 24-24 24zm0-45C12.4 3 3 12.4 3 24s9.4 21 21 21 21-9.4 21-21S35.6 3 24 3z"></path><path d="M34.9 24c0-1.4-1.1-2.5-2.5-2.5s-2.5 1.1-2.5 2.5 1.1 2.5 2.5 2.5 2.5-1.1 2.5-2.5zm-21.8 0c0-1.4 1.1-2.5 2.5-2.5s2.5 1.1 2.5 2.5-1.1 2.5-2.5 2.5-2.5-1.1-2.5-2.5zM24 37.3c-5.2 0-8-3.5-8.2-3.7-.5-.6-.4-1.6.2-2.1.6-.5 1.6-.4 2.1.2.1.1 2.1 2.5 5.8 2.5 3.7 0 5.8-2.5 5.8-2.5.5-.6 1.5-.7 2.1-.2.6.5.7 1.5.2 2.1 0 .2-2.8 3.7-8 3.7z"></path></svg>
       		<input id={`CB${props.postId}`} type="text" value={singleComment} onChange={(e) => { setSingleComment(e.target.value) }} placeholder="Add a comment..."/>
-      		<a href="#!" onClick={handlePostComment}>Post</a>
+      		<a href="#!" onClick={handlePostComment} className={(singleComment !== "")?(""):("disable")}>Post</a>
       	</div>
       </div>
     </div>
